Move API related files to a subfolder
