+++ monlog/log/api/api.py
@@ -5,10 +5,10 @@ from django.db.models import Q
 from tastypie import fields
 from tastypie.resources import ModelResource, ALL
 from tastypie.authorization import DjangoAuthorization
-from log.authentication import MonlogAuthentication
-from log.authentication import CookieAuthentication
+from log.api.authentication import MonlogAuthentication
+from log.api.authentication import CookieAuthentication
 from log.models import LogMessage, SEVERITY_CHOICES
-from log.validation import LogValidation
+from log.api.validation import LogValidation
 from datetime import datetime 
 
 class ApplicationResource(ModelResource):
